[COT-217] Refactor: 사용하지 않는 API 및 DTO 제거 (#340)
* refactor: remove deprecated api

* refactor: move directory
