--- src/main/java/org/cotato/csquiz/api/admin/dto/MemberEnrollInfoResponse.java
@@ -1,25 +1,0 @@
-package org.cotato.csquiz.api.admin.dto;
-
-import org.cotato.csquiz.domain.auth.enums.MemberPosition;
-import org.cotato.csquiz.domain.auth.enums.MemberRole;
-import org.cotato.csquiz.domain.auth.entity.Member;
-
-public record MemberEnrollInfoResponse(
-        Long memberId,
-        String name,
-        MemberPosition position,
-        Integer generationNumber,
-        MemberRole role
-) {
-
-
-    public static MemberEnrollInfoResponse of(Member member) {
-        return new MemberEnrollInfoResponse(
-                member.getId(),
-                member.getName(),
-                member.getPosition(),
-                member.getPassedGenerationNumber(),
-                member.getRole()
-        );
-    }
-}
