--- src/main/java/org/cotato/csquiz/api/admin/dto/MemberRejectRequest.java
@@ -1,9 +1,0 @@
-package org.cotato.csquiz.api.admin.dto;
-
-import jakarta.validation.constraints.NotNull;
-
-public record MemberRejectRequest(
-        @NotNull
-        Long memberId
-) {
-}
