+++ src/main/java/org/cotato/csquiz/api/member/controller/MemberController.java
@@ -7,10 +7,10 @@ import java.io.IOException;
 import javax.naming.NoPermissionException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.cotato.csquiz.api.admin.dto.MemberApproveRequest;
-import org.cotato.csquiz.api.admin.dto.MemberInfoResponse;
-import org.cotato.csquiz.api.admin.dto.UpdateActiveMemberToOldMemberRequest;
-import org.cotato.csquiz.api.admin.dto.UpdateMemberRoleRequest;
+import org.cotato.csquiz.api.member.dto.MemberApproveRequest;
+import org.cotato.csquiz.api.member.dto.MemberInfoResponse;
+import org.cotato.csquiz.api.member.dto.UpdateActiveMemberToOldMemberRequest;
+import org.cotato.csquiz.api.member.dto.UpdateMemberRoleRequest;
 import org.cotato.csquiz.api.member.dto.DeactivateRequest;
 import org.cotato.csquiz.api.member.dto.MemberMyPageInfoResponse;
 import org.cotato.csquiz.api.member.dto.MemberResponse;
